[fix] specialchar in social-graph metadata
It appears that decoding the HTML entities with `html_entity_decode()` 
was a non-goal as it is already managed by WordPress in almost all the 
case. The escaping of HTML is still needed when getting the post meta. 
`wp_strip_all_tags` is needed in case we use an HTML excerpt or 
description.

This fixes #39